Move bot config constants & re-implement
