+++ src/Constants/BotConfig/EmbedColors.ts
